Add initial Date#_strptime support
Much like parse and _parse, the work in strptime needs to be done
by _strptime so Time.strptime can work.

Add rhrd__from_hash to create a Date object from a hash of values,
so that the code from parse can be used in strptime.  This is
the equivalent of new_from_frags in the stdlib.