Add DateTime#<=>
I don't really like the implementation, but checking for equality
when using floating point is always a pain.  In this case, we
consider anything within a millisecond as being equal.

Fix todo and toofar methods.  The underlying methods checked
needed to be changed.

While here, make sure DateTime.today is not defined.