Make DateTime#ajd and amjd accurate
Previously, they just used jd and mjd.  For Date I think that is
acceptable, since a date is technically a 24 hour time period,
and the ajd and amjd fall inside the range of time for a given
date.  However, since DateTime is an exact value, it should be
return the actual value.