Various fixes for better spec compatibility
Correctly handle the %s strftime modifier to include seconds in
the current day.

Make Date#start return a number lower than the lowest possible
date, instead of higher than the highest possible date, as we
always consider the gregorian start time in the past.

Remove the use of Rational in the pure ruby parsing code.

Remove a debug printf and fix the handling of the %s strptime
modifier for DateTime.

Handle an offset but no hour, minute, or second in
DateTime.strptime.

Populate the jd and nanos fields when creating a datetime with
a new offset.

Consider offsets when comparing DateTimes to each other in
eql? and <=>.  However, don't consider them when comparing
a DateTime to Date or Numeric.

Fix DateTime#<=> Numeric when the day matches but the times
do not.