Use at least 2 seconds for all benchmarks for greater accuracy
This removes the bench_quick and bench_long tasks, and uses at
least 2 seconds for all benchmarks.  It runs the benchmarks in
batches of 10,000 iterations, repeating if the time elapsed is
less than 2 seconds.  The report changes from the total time
elapsed to the average time in nanoseconds for each call.