Give extension a shot at working for compiled extensions not ending in .so
Instead of hard coding date.so, rename the extension from
date.c to date_ext.c, and just require date_ext. Most of
the other changes are fallout from that.

The windows_gem task is now more thorough.  I'm not sure
why in some cases you need to manually copy the files,
but handle that.