Tighten up offset limits
Make the 14 hour offset limit hard, so 14 hours and 1 minute is
considered over the limit.  To do this, always convert an offset as
day fraction to offset in minutes before doing the range check.

DRY up some code by adding rhrdt__check_offset to check the offset
and raise, and rename some __valid_* functions to __set_* and have
them raise on error instead of returning 0/1.