Minor cleanup of the Ragel parser
Since the plan is to support multiple formats, move the setting of
the parser type used to an EOF action of the specific parser type.
Remove the general EOF action for all types, and move the main
switch statement to after the parsing.  This fixes an unlikely but
possible issue if one parser would match partially but not
completely, and another parser would match completely.

While here, allow a lowercase t to separate the ISO date from time.