Fix the :zone entry when using _ragel_parse
This adds a t_iso_zone_end pointer to store the end of the string.
Previously, this wasn't necessary, as the end of the iso_date_time
was the end of the string.  However, since there can now be spaces
at the end, the end needs to be kept track of separately.