Try to fix some issues where loading date/format directly breaks
This should fix an issue if you attempt to load date/format
directly (or indirectly by requiring time or parsedate without
requiring date).  Before, it just checked if Date was defined),
but it's possible someone else may have opened (and hence
defined the class).  Instead, check for Date::ZONES, which is
unlikely to be defined by anyone.

Note that this can still fail if date has already been required,
but for some reason it was the standard lib version, or if
it was home_run's version, but the Date constant was later
removed.