fix free function for CALC_EXACT_MALLOC_SIZE
When Ruby (1.9) is built with CALC_EXACT_MALLOC_SIZE, it will
allocate an extra word to track the size of each allocation.
This throws off the use of free() on memory allocated by
(vm_)xmalloc.  Just pass "-1" instead as Ruby will
automatically call the correct free function (xfree()) as
documented in README.EXT.