Separately compile each C file, instead of including C files in other C files
Switch to the standard C way of doing things, with a shared header
file.  This bloats the code a bit as we have to add prototypes that
were unnecessary before, as well as add extern for some static
variables.  However, it's probably better to do things in a standard
way.  Thanks to Luis Lavena for suggesting this.