Fix DateTime#sec_fraction on ruby 1.9
Previously, it had the 1.8 behavior of returning the fraction of
the second as a fraction of the day, even on 1.9.  This increases
compatibility with the 1.9 stdlib.

This also adds a spec for the :sec_fraction entry of Date._parse,
though that is the fraction of the second on both 1.8 and 1.9.