Add specs for subclass methods returning subclass instances
Previously, Date was hard coded in a few places, so that if you
did Class.new(Date).new + 1, you got an instance of Date back,
instead of an instance of the subclass. This adds specs for many
cases to ensure that the correct class is used.  In some cases,
these specs pass without code changes, in others the following
commit is required.

Problem pointed out by aquasync (ruby-ole gem author) as a comment
on issue #15.