Define Date::Format::ZONES in ruby instead of C, bump version to 1.0.0
For some unknown reason, defining Date::Format::ZONES in C was
causing the occassional segfault when loading home_run, either
in an rb_hash_aset call or when freezing the hash in format.rb.

The reason it was defined in C originally was because it is used
in the zone_to_diff method defined in C.  However, that can be
worked around with some additionally checking in the method.

Note that Date::ZONES is no longer defined, only Date::Format::ZONES.