Enable allocate class method for Date and DateTime
Also, change dup and clone to call super instead of using
DUPSETUP and CLONESETUP.  This makes dup and clone work
correctly on rubinius.  It also makes dup and clone work
correctly for Date on jruby, but not for DateTime due to
a bug in JRuby.

allocate is defined to use the julian day 0 for simplicity.